More CUDD functions; start of higher-level interface & tests
The high-level interface provides a single, implicit manager for each BDD
computation.  A BDD computation is parameterized by a phantom type variable,
and a rank-2 type for BDD computations & BDD references is used to ensure that
no BDD references escapes from its manager.

This perhaps isn't the best way to create a higher-level interface: it
precludes simultaneous use of multiple managers, and the interface is a bit
complicated due to the types.  However, I am working under the maxim that it
should not be possible to write a Haskell program that segfaults; this
high-level interface I've described should prevent several kinds of CUDD
errors.

This high-level interface currently does nothing with BDD reference counts.
This is a problem.