CUDD garbage collection now triggers Haskell garbage collection.
By collecting Haskell garbage before collecting CUDD garbage, Haskell
references to BDDs may be freed, allowing CUDD to collect the corresponding
nodes.

There is a problem with this implementation:  each Mgr created with newMgr will
leak a bit of memory for all the hook functions that are registered.  It is not
obvious to me how to avoid this.