PropHC -> BDD evaluation, plus reverse breadth-first traveral.
The hope is that through common subexpression elimination, followed
by scheduling of the subcomponents for evaluation in lowest-first
scheduling (i.e., evaluating upward from the leaves of the tree,
which is different than postorder traversal), that BDD synthesis
would be faster than by simply doing a postorder evaluation of the
raw (non-CSEd) sentences of propositional logic.

I first thought that a reversed breadth-first traversal of the
subcomponents of the CSEd sentence would give me the lowest-first
schedule that I want.  However, because I only visit each node once,
the reversed breadth-first traversal doesn't work in general.  (See
the property tests for examples.)

Also, I don't know if the CSE of propositional sentences will be
pointless with BDD synthesis anyway--re-evaluation of subcomponents
of the sentence should turn into some hash table lookups of the
already-computed results in the BDD data structure.